+++ client/src/pages/joinsession.js
